feat: show message if nothing to be rerun
If the journal indicates that all contract deploys/calls have already
completed, then display a message to that effect at the cli and then
display the previously deployed contract addresses.

Supporting refactor:

* refactor(core): merge batch into execution

In the short term there is no other planned execution policy other than
batching so merge the batching code into execution proper.
