+++ packages/core/test/execution/batching.ts
@@ -4,14 +4,14 @@ import { BigNumber } from "ethers";
 
 import { Deployment } from "deployment/Deployment";
 import { ExecutionGraph } from "execution/ExecutionGraph";
-import { visitInBatches } from "execution/batch/visitInBatches";
+import { executeInBatches } from "execution/execute";
 import { ContractDeploy, ExecutionVertex } from "types/executionGraph";
 import { VertexVisitResult } from "types/graph";
 import { ICommandJournal } from "types/journal";
 
 import { buildAdjacencyListFrom } from "../graph/helpers";
 
-describe("Execution - visitInBatches", () => {
+describe("Execution - batching", () => {
   it("should run", async () => {
     const vertex0: ExecutionVertex = createFakeContractDeployVertex(0, "first");
     const vertex1: ExecutionVertex = createFakeContractDeployVertex(
@@ -32,7 +32,10 @@ describe("Execution - visitInBatches", () => {
     executionGraph.vertexes.set(2, vertex2);
 
     const mockServices = {} as any;
-    const mockJournal: ICommandJournal = { record: async () => {} };
+    const mockJournal: ICommandJournal = {
+      record: async () => {},
+      read: () => null,
+    };
     const mockUpdateUiAction = () => {};
 
     const deployment = new Deployment(
@@ -42,7 +45,7 @@ describe("Execution - visitInBatches", () => {
       mockUpdateUiAction
     );
 
-    const result = await visitInBatches(
+    const result = await executeInBatches(
       deployment,
       executionGraph,
       async (): Promise<VertexVisitResult> => {
@@ -51,6 +54,7 @@ describe("Execution - visitInBatches", () => {
       {
         maxRetries: 4,
         gasIncrementPerRetry: null,
+        pollingInterval: 300,
       }
     );
